Dimension-unified the wavefield helper struct.
