+++ core/specfem/assembly/assembly/dim3/compute_wavefield/compute_wavefield.cpp
@@ -1,6 +1,6 @@
 #include "enumerations/interface.hpp"
-#include "helper.hpp"
 #include "specfem/assembly/assembly.hpp"
+#include "specfem/assembly/assembly/impl/helper.hpp"
 #include "specfem/macros.hpp"
 #include <Kokkos_Core.hpp>
 #include <stdexcept>
